requirements.txt
paho-mqtt==1.5.0
schedule~=0.6.0